ml/requirements.txt
# ML Dependencies for Genre Classifier

# Data processing
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0

# ML/NLP
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0

# API serving
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
